add error handling

+++ netlify/functions/deleteEvent.js
@@ -1,18 +1,26 @@
-const {Event} = require("../../Database/Models");
+const { Event } = require("../../Database/Models");
 export async function handler(event) {
   const headers = {
-    'Access-Control-Allow-Origin': '*',
-    'Access-Control-Allow-Headers': '*',
-    'Access-Control-Allow-Methods': '*'
+    "Access-Control-Allow-Origin": "*",
+    "Access-Control-Allow-Headers": "*",
+    "Access-Control-Allow-Methods": "*",
   };
-  const deletedEvent = await Event.destroy({
-     where : {
-       id : event.body.id,
-     }
-   });
-  return {
-    statusCode : 200,
-    headers,
-    body : JSON.stringify(deletedEvent)
+  if (event.httpMethod === "OPTIONS") {
+    return {
+      statusCode: 200,
+      headers,
+      body: "Successful preflight call.",
+    };
+  } else if (event.httpMethod === "DELETE") {
+    const deletedEvent = await Event.destroy({
+      where: {
+        id: event.body.id,
+      },
+    });
+    return {
+      statusCode: 200,
+      headers,
+      body: JSON.stringify(deletedEvent),
+    };
   }
 }
